+++ spec/ffi/rbx/memory_pointer_spec.rb
@@ -130,13 +130,11 @@ describe "MemoryPointer" do
   end
 
   it "raises an error if you try putting a long into a pointer of size 1" do
-    skip "not yet supported on TruffleRuby" if RUBY_ENGINE == "truffleruby"
     m = FFI::MemoryPointer.new(1)
     expect { m.write_long(10) }.to raise_error(IndexError)
   end
 
   it "raises an error if you try putting an int into a pointer of size 1" do
-    skip "not yet supported on TruffleRuby" if RUBY_ENGINE == "truffleruby"
     m = FFI::MemoryPointer.new(1)
     expect { m.write_int(10) }.to raise_error(IndexError)
   end