+++ spec/ruby/core/kernel/shared/require.rb
@@ -351,9 +351,13 @@ describe :kernel_require, shared: true do
 
           symlink_to_code_dir_two = tmp("codesymlinktwo")
           File.symlink("#{CODE_LOADING_DIR}/b", symlink_to_code_dir_two)
-          $LOAD_PATH.unshift(symlink_to_code_dir_two)
+          begin
+            $LOAD_PATH.unshift(symlink_to_code_dir_two)
 
-          @object.require('load_fixture').should be_false
+            @object.require('load_fixture').should be_false
+          ensure
+            rm_r symlink_to_code_dir_two
+          end
         end
       end